Added Playlist feature
Functions were added in the spotifyAPI for the playlist feature to work once the user is connected to it. Playlist logic should be set and ready, add to playlist button was also added as well as a drop down menu for it for the user to select the playlist. UI still needs to be updated as some tailwind code or css code adjusted the home page and size of the playback card.